requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.2.1
sqlalchemy==2.0.30
aiosqlite==0.20.0
slowapi==0.1.9
structlog==24.1.0
httpx==0.27.0
pytest==8.2.0
pytest-asyncio==0.23.6
